updating for simpler check
