+++ spec/javascript/e2e/ApplicationEditPage.e2e.js
@@ -1,21 +1,8 @@
-import puppeteer from 'puppeteer'
-
-import sharedSteps from '../support/puppeteer/steps/sharedSteps'
-import { DEFAULT_E2E_TIME_OUT, HEADLESS, LEASE_UP_LISTING_APPLICATION_ID } from '../support/puppeteer/consts'
+import { applicationRouteCheck } from '../support/puppeteer/steps/applications'
+import { DEFAULT_E2E_TIME_OUT, LEASE_UP_LISTING_APPLICATION_ID } from '../support/puppeteer/consts'
 
 describe('ApplicationEditPage', () => {
-  test('should redirect when lottery_status is anything other than "Not Yet Run"', async () => {
-    let browser = await puppeteer.launch({ headless: HEADLESS })
-    let page = await browser.newPage()
-
-    await sharedSteps.loginAsAgent(page)
-    await sharedSteps.goto(page, `/applications/${LEASE_UP_LISTING_APPLICATION_ID}/edit`)
-    await page.waitFor(2000)
-
-    const currentUrl = page.url()
-    expect(currentUrl).toContain(`http://localhost:3000/applications/${LEASE_UP_LISTING_APPLICATION_ID}`)
-    expect(currentUrl).not.toContain(`edit`)
-
-    await browser.close()
+  test('should redirect edit application when lottery_status is anything other than "Not Yet Run"', async () => {
+    await applicationRouteCheck('edit', LEASE_UP_LISTING_APPLICATION_ID)
   }, DEFAULT_E2E_TIME_OUT)
 })
