+++ spec/javascript/e2e/ApplicationNewPage.e2e.js
@@ -2,6 +2,7 @@ import puppeteer from 'puppeteer'
 
 import utils from '../support/puppeteer/utils'
 import sharedSteps from '../support/puppeteer/steps/sharedSteps'
+import { applicationRouteCheck } from '../support/puppeteer/steps/applications'
 import {
   NON_LEASE_UP_LISTING_ID, DEFAULT_E2E_TIME_OUT, HEADLESS, SALE_LISTING_ID, LEASE_UP_LISTING_ID,
   FIRST_NAME, LAST_NAME, TRUNCATED_FIRST_NAME, TRUNCATED_LAST_NAME, DOB_MONTH, DOB_DAY, DOB_YEAR,
@@ -323,17 +324,6 @@ describe('ApplicationNewPage', () => {
   }, DEFAULT_E2E_TIME_OUT)
 
   test('should redirect when lottery_status is anything other than "Not Yet Run"', async () => {
-    let browser = await puppeteer.launch({ headless: HEADLESS })
-    let page = await browser.newPage()
-
-    await sharedSteps.loginAsAgent(page)
-    await sharedSteps.goto(page, `/listings/${LEASE_UP_LISTING_ID}/applications/new`)
-    await page.waitFor(2000)
-
-    const currentUrl = page.url()
-    expect(currentUrl).toContain(`http://localhost:3000/listings/${LEASE_UP_LISTING_ID}`)
-    expect(currentUrl).not.toContain(`applications/new`)
-
-    await browser.close()
+    await applicationRouteCheck('new', LEASE_UP_LISTING_ID)
   }, DEFAULT_E2E_TIME_OUT)
 })
